+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added travis CI integration
 - Removed callbacks still kicking around in the couch notifier.
+- Renamed state to checkpoint as it is more accurate
 
 ## 0.4.4